Pub: Extract PubCacheReader to a separate file
Reduce the size of the large `Pub.kt` a bit.
